# PUBDEV-5170: Expose individual predictions of GBM/DRF trees in MOJO API