--- a/h2o-algos/src/jmh/resources/hex/tree/gbm/prostate/model.ini
+++ b/h2o-algos/src/jmh/resources/hex/tree/gbm/prostate/model.ini
@@ -1,0 +1,38 @@
+[info]
+h2o_version = 3.21.0.99999
+mojo_version = 1.20
+license = Apache License Version 2.0
+algo = gbm
+algorithm = Gradient Boosting Machine
+endianness = LITTLE_ENDIAN
+category = Binomial
+uuid = -1252995762865247240
+supervised = true
+n_features = 7
+n_classes = 2
+n_columns = 8
+n_domains = 2
+balance_classes = false
+default_threshold = 0.4676837164777207
+prior_class_distrib = [0.5973684210526315, 0.4026315789473684]
+model_class_distrib = [0.5973684210526315, 0.4026315789473684]
+timestamp = 2018-06-25T15:10:25.169-07:00
+n_trees = 50
+n_trees_per_class = 1
+distribution = bernoulli
+init_f = -0.3945120960889672
+offset_column = null
+
+[columns]
+AGE
+RACE
+DPROS
+DCAPS
+PSA
+VOL
+GLEASON
+CAPSULE
+
+[domains]
+1: 3 d000.txt
+7: 2 d001.txt