Scoring performance is not affected by this change.

```
== NEW ==

Result "hex.tree.gbm.GbmMojoScoringBench.measureGbmScore0":
  5.636 ±(99.9%) 0.060 ms/op [Average]
  (min, avg, max) = (5.560, 5.636, 5.681), stdev = 0.040
  CI (99.9%): [5.576, 5.696] (assumes normal distribution)


# JMH version: 1.21
# VM version: JDK 1.7.0_80, Java HotSpot(TM) 64-Bit Server VM, 24.80-b11
# VM invoker: /Library/Java/JavaVirtualMachines/jdk1.7.0_80.jdk/Contents/Home/jre/bin/java
# VM options: <none>
# Warmup: 3 iterations, 10 s each
# Measurement: 10 iterations, 10 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: hex.tree.gbm.GbmMojoScoringBench.measureGbmScore0
# Parameters: (rows = 100000)

# Run progress: 50.00% complete, ETA 00:02:10
# Fork: 1 of 1

# Warmup Iteration   1: 573.318 ms/op
# Warmup Iteration   2: 559.469 ms/op]
# Warmup Iteration   3: 562.324 ms/op]
Iteration   1: 552.601 ms/opG [3m 1s]
Iteration   2: 559.078 ms/opG [3m 11s]
Iteration   3: 554.794 ms/opG [3m 21s]
Iteration   4: 559.212 ms/opG [3m 31s]
Iteration   5: 564.856 ms/opG [3m 41s]
Iteration   6: 561.050 ms/opG [3m 51s]
Iteration   7: 565.216 ms/opG [4m 2s]
Iteration   8: 556.939 ms/opG [4m 12s]
Iteration   9: 565.799 ms/opG [4m 22s]
Iteration  10: 554.681 ms/opG [4m 32s]

> Task :h2o-algos:jmh


Result "hex.tree.gbm.GbmMojoScoringBench.measureGbmScore0":
  559.422 ±(99.9%) 7.183 ms/op [Average]
  (min, avg, max) = (552.601, 559.422, 565.799), stdev = 4.751
  CI (99.9%): [552.239, 566.606] (assumes normal distribution)

== OLD ==

# JMH version: 1.21
# VM version: JDK 1.7.0_80, Java HotSpot(TM) 64-Bit Server VM, 24.80-b11
# VM invoker: /Library/Java/JavaVirtualMachines/jdk1.7.0_80.jdk/Contents/Home/jre/bin/java
# VM options: <none>
# Warmup: 3 iterations, 10 s each
# Measurement: 10 iterations, 10 s each
# Timeout: 10 min per iteration
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: hex.tree.gbm.GbmMojoScoringBench.measureGbmScore0
# Parameters: (rows = 100000)

# Run progress: 50.00% complete, ETA 00:02:10
# Fork: 1 of 1

# Warmup Iteration   1: 568.800 ms/op
# Warmup Iteration   2: 558.028 ms/op]
# Warmup Iteration   3: 556.847 ms/op]
Iteration   1: 559.942 ms/opG [3m 3s]
Iteration   2: 557.310 ms/opG [3m 13s]
Iteration   3: 565.237 ms/opG [3m 24s]
Iteration   4: 551.948 ms/opG [3m 34s]
Iteration   5: 559.610 ms/opG [3m 44s]
Iteration   6: 554.633 ms/opG [3m 54s]
Iteration   7: 556.770 ms/opG [4m 4s]
Iteration   8: 562.597 ms/opG [4m 14s]
Iteration   9: 581.520 ms/opG [4m 25s]
Iteration  10: 555.411 ms/opG [4m 35s]

> Task :h2o-algos:jmh


Result "hex.tree.gbm.GbmMojoScoringBench.measureGbmScore0":
  560.498 ±(99.9%) 12.622 ms/op [Average]
  (min, avg, max) = (551.948, 560.498, 581.520), stdev = 8.349
  CI (99.9%): [547.875, 573.120] (assumes normal distribution)


# Run complete. Total time: 00:04:23

Benchmark                             (rows)  Mode  Cnt    Score    Error  Units
GbmMojoScoringBench.measureGbmScore0    1000  avgt   10    5.701 ±  0.352  ms/op
GbmMojoScoringBench.measureGbmScore0  100000  avgt   10  560.498 ± 12.622  ms/op
```